requirements.txt
# ── Web サーバ ──
Flask==2.3.2
gunicorn==23.0.0

# ── Google Cloud ──
google-auth>=2.20.0
google-api-python-client>=2.70.0
google-cloud-firestore>=2.20.0

# ── OpenAI ──
openai>=1.3.0   # chat + whisper API 両対応
requests>=2.31.0


